+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * KLS-437 - Change DIT references to DBT
 * KLS-407 - Add format filters to the Export Academy event listings page
 * KLS-270 - Add webinar duration to event listing template
+* KLS-426 - Add event types to filterset on UKEA's listings page
 
 ## [2.27.0](https://github.com/uktrade/great-cms/releases/tag/2.27.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/2.26.0...2.27.0)